refactor(imports): update import paths to use gameEnums for consistency and improved organization
