+++ client/src/components/actionBar/ActionBar.module.css
@@ -1,4 +1,3 @@
-/* Sticker component styles */
 .actionBarContainer {
   width: 100%;
   height: 15%;
